requirements.txt
Django==1.8
django-angular==0.7.13
django-debug-toolbar==1.3.0
django-filter==0.9.2
django-widget-tweaks==1.3
djangorestframework==3.1.1
MySQL-python==1.2.5
simplejson==3.6.5
sqlparse==0.1.15
uWSGI==2.0.10
